face_rec/requirements.txt
click==8.2.1
cmake==4.0.2
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
PyMySQL==1.1.1
setuptools==80.9.0
